myapp/tavern_of_heroes/api/serializers.py
from rest_framework import serializers
from ..models import Item, Hero


class HeroSerializer(serializers.ModelSerializer):
    class Meta:
        model = Hero
        fields = '__all__'


class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = '__all__'
